dev: Refactor locale settings into their own file.
